[DEP-103] split up testData, organize, cleanup, env dist update
