+++ .env.dist
@@ -13,7 +13,11 @@ AUTH0_CLIENT_ID=None
 AUTH0_CLIENT_SECRET=None
 AUTH0_AUDIENCE=https://localhost:8000
 CI=false
-MINIMAL_SEED=true
+
+# set to true if desired for local builds
+# will seed only necessary data for app functionality
+# dates, ticket types, discounts etc. See seed.ts in server/prisma
+MINIMAL_SEED=false
 
 # For Playwright E2E Testing
 TEST_EMAIL=none
