+++ .github/workflows/E2E_tests.yml
@@ -67,8 +67,8 @@ jobs:
       - name: docker setup
         run: docker-compose up -d
 
-      - name: 20 seconds wait for client and server start, database seed
-        run: sleep 20s
+      - name: 30 seconds wait for client and server start, database seed
+        run: sleep 30s
 
       - name: Check client E2E(UI) Tests
         run: npx playwright test
