+++ client/e2e/pages/loginPage.ts
@@ -1,4 +1,6 @@
-import { type Locator, type Page } from '@playwright/test';
+/* eslint-disable require-jsdoc */
+import {type Locator, type Page } from '@playwright/test';
+
 
 export class LoginPage {
   readonly page: Page;
