+++ client/e2e/pages/mainPage.ts
@@ -1,7 +1,6 @@
 /* eslint-disable require-jsdoc */
-
 import {type Locator, type Page, expect} from '@playwright/test';
-import {EventInfo, CreditCard, Customer} from '../testData/ConstsPackage';
+import { CreditCardInfo, CustomerInfo, EventInfo } from '../test-data/interfaces';
 
 export class MainPage {
   readonly page: Page;
@@ -236,7 +235,7 @@ export class MainPage {
   }
 
   // Fills out the customer info from the customer parameter
-  async fillCustomerInfo(customer: Customer) {
+  async fillCustomerInfo(customer: CustomerInfo) {
     await this.cartFirstName.fill(customer.firstName);
     await this.cartLastName.fill(customer.lastName);
     await this.cartStreetAddress.fill(customer.streetAddress);
@@ -257,7 +256,7 @@ export class MainPage {
   // Fill out data on Stripe page.  Currently uses both a Customer and CreditCard.
   // Stripe is slow and sometimes has an account popup after email entry.
   // This function waits to see if it will pop up and handle it appropriately.
-  async fillStripeInfo(customer: Customer, ccInfo: CreditCard, timeoutAdd = 0) {
+  async fillStripeInfo(customer: CustomerInfo, ccInfo: CreditCardInfo, timeoutAdd = 0) {
     await this.stripeEmail.fill(customer.email);
     await this.page.waitForTimeout(10000 + timeoutAdd);
     if (await this.page.getByText('Use your saved information').isVisible()) {
@@ -285,8 +284,8 @@ export class MainPage {
   // Date, time, and ticket type will be selected randomly.
   // Qty will be used to determine quantity if passed, otherwise will use a random quantity as well.
   async purchaseTicket(
-    customer: Customer,
-    creditCard: CreditCard,
+    customer: CustomerInfo,
+    creditCard: CreditCardInfo,
     event: EventInfo,
     options?: {qty?: number; timeoutAdd?: number},
   ) {
