+++ client/e2e/test-data/dataConstants/CreditCardConstants.ts
@@ -1,0 +1,8 @@
+import { CreditCardInfo } from "../interfaces";
+
+export const VALID_VISA_CREDIT: CreditCardInfo = {
+  cardNumber: '4242 4242 4242 4242',
+  CVC: '999',
+  date: '12 / 30',
+};
+
