+++ client/e2e/test-data/dataConstants/EventInfoConstants.ts
@@ -1,0 +1,56 @@
+import { EventInfo } from "../interfaces";
+
+export const EVENT_INFO_1: EventInfo = {
+  eventName: 'S',
+  eventDescription: '123',
+  eventURL: 'http://',
+  eventFullName: 'S Playbill S Description 123',
+};
+
+export const EVENT_INFO_2: EventInfo = {
+  eventName: 'Test_event',
+  eventDescription: 'An event for testing',
+  eventURL:
+    'https://www.hindustantimes.com/ht-img/img/2023/08/25/550x309/international_dog_day_1692974397743_1692974414085.jpg',
+  eventFullName:
+    'Test_event Playbill Test_event Description An event for testing',
+};
+
+export const EVENT_INFO_3: EventInfo = {
+  eventName: 'The Crucible1',
+  eventDescription:
+    '111Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+  eventURL:
+    'https://www.hindustantimes.com/ht-img/img/2023/08/25/550x309/international_dog_day_1692974397743_1692974414085.jpg',
+  eventFullName:
+    'The Crucible1 Playbill The Crucible1 Description 111Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+};
+
+export const EVENT_INFO_4: EventInfo = {
+  eventName: 'The Crucible',
+  eventDescription:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+  eventURL: 'https://upload.wikimedia.org/wikipedia/en/7/75/Cruciblecover.jpg',
+  eventFullName:
+    'The Crucible Playbill The Crucible Description Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+};
+
+export const EVENT_INFO_5: EventInfo = {
+  eventName: 'The Martyr',
+  eventDescription:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+  eventURL:
+    'https://upload.wikimedia.org/wikipedia/commons/9/9e/Martirio_de_San_Esteban_%28Correa_de_Vivar%29.jpg',
+  eventFullName:
+    'The Martyr Playbill The Martyr Description Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+};
+
+export const EVENT_INFO_6: EventInfo = {
+  eventName: 'Purple',
+  eventDescription:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+  eventURL:
+    'https://upload.wikimedia.org/wikipedia/en/6/62/Barney%27s-Great-Adventure-Poster.jpeg',
+  eventFullName:
+    'Purple Playbill Purple Description Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+};
