+++ client/e2e/test-data/dataConstants/SeasonInfoConstants.ts
@@ -1,0 +1,17 @@
+import { SeasonInfo } from "../interfaces";
+
+export const SEASON_INFO_1: SeasonInfo = {
+  seasonName: 'Test_Season1',
+  seasonStart: '2020-02-02',
+  seasonEnd: '2020-03-02',
+  seasonImgURL:
+    'https://www.hindustantimes.com/ht-img/img/2023/08/25/550x309/international_dog_day_1692974397743_1692974414085.jpg',
+};
+
+export const SEASON_INFO_2: SeasonInfo = {
+  seasonName: 'Test_Season2',
+  seasonStart: '2021-02-02',
+  seasonEnd: '2021-03-02',
+  seasonImgURL:
+    'https://www.hindustantimes.com/ht-img/img/2023/08/25/550x309/international_dog_day_1692974397743_1692974414085.jpg',
+};
