+++ client/e2e/tests/doorListTest.spec.ts
@@ -1,9 +1,12 @@
-import {test, expect} from '@playwright/test';
-import {DoorListPage} from './pages/doorListPage';
-import {EventsPage} from './pages/EventsPage';
-import {EventInfo, EventInfoTemplate2, ShowingInfo1, ValidVisaCredit} from './testData/ConstsPackage';
-import { baseJaneDoe, createUniqueCustomer } from './testData/ConstsPackage';
-import { MainPage } from './pages/mainPage';
+import { test, expect } from '@playwright/test';
+import { DoorListPage } from '../pages/doorListPage';
+import { EventsPage } from '../pages/EventsPage';
+import { createUniqueCustomer, createUniqueEvent } from '../test-data/factoryFunctions';
+import { EVENT_INFO_2 } from '../test-data/dataConstants/EventInfoConstants';
+import { MainPage } from '../pages/mainPage';
+import { SHOWING_INFO_1 } from '../test-data/dataConstants/ShowingInfoConstants';
+import { JANE_DOE } from '../test-data/dataConstants/CustomerInfoConstants';
+import { VALID_VISA_CREDIT } from '../test-data/dataConstants/CreditCardConstants';
 
 
 test('Check Home', async ({page}) => {
@@ -13,19 +16,20 @@ test('Check Home', async ({page}) => {
 });
 
 test('Select Active Showing in Doorlist', async ({page}) => {
-  const currentEvent = new EventInfo(EventInfoTemplate2);
+  const currentEvent = createUniqueEvent(EVENT_INFO_2);
   const eventsPage = new EventsPage(page);
   const doorList = new DoorListPage(page);
+
   try {
     // Add event and showing to check for in the door list
     await eventsPage.goto();
     await eventsPage.addnewevent(currentEvent);
     await eventsPage.activateEvent();
-    await eventsPage.addNewShowing(ShowingInfo1);
+    await eventsPage.addNewShowing(SHOWING_INFO_1);
 
     // Check door list
     await doorList.goto();
-    await doorList.searchShowing(currentEvent,ShowingInfo1);
+    await doorList.searchShowing(currentEvent, SHOWING_INFO_1);
   } finally {
     // Remove the added event
     await eventsPage.goto();
@@ -35,19 +39,20 @@ test('Select Active Showing in Doorlist', async ({page}) => {
 });
 
 test('Select Inactive Showing in Doorlist', async ({page}) => {
-  const currentEvent = new EventInfo(EventInfoTemplate2);
+  const currentEvent = createUniqueEvent(EVENT_INFO_2);
   const eventsPage = new EventsPage(page);
   const doorList = new DoorListPage(page);
+
   try {
     // Add event and showing to check for in the door list
     await eventsPage.goto();
     await eventsPage.addnewevent(currentEvent);
-    await eventsPage.addNewShowing(ShowingInfo1);
+    await eventsPage.addNewShowing(SHOWING_INFO_1);
 
     // Check door list
     await doorList.goto();
     await doorList.setAllView()
-    await doorList.searchShowing(currentEvent,ShowingInfo1);
+    await doorList.searchShowing(currentEvent,SHOWING_INFO_1);
   } finally {
     // Remove the added event
     await eventsPage.goto();
@@ -58,8 +63,8 @@ test('Select Inactive Showing in Doorlist', async ({page}) => {
 });
 
 test('Open Seats in Doorlist', async ({page}) => {
-  const currentEvent = new EventInfo(EventInfoTemplate2);
-  const currentShowing = ShowingInfo1;
+  const currentEvent = createUniqueEvent(EVENT_INFO_2);
+  const currentShowing = SHOWING_INFO_1;
   const eventsPage = new EventsPage(page);
   const doorList = new DoorListPage(page);
   try {
@@ -85,10 +90,10 @@ test('Open Seats in Doorlist', async ({page}) => {
 test('Purchased Seats in Doorlist', async ({page}, testInfo) => {
   const timeoutAdd = testInfo.retry * 5000;
   test.setTimeout(80000 + timeoutAdd);
-  const currentEvent = new EventInfo(EventInfoTemplate2);
-  const currentShowing = ShowingInfo1;
-  const currentPatron = createUniqueCustomer(baseJaneDoe);
-  const currentCard = ValidVisaCredit;
+  const currentEvent = createUniqueEvent(EVENT_INFO_2);
+  const currentShowing = SHOWING_INFO_1;
+  const currentPatron = createUniqueCustomer(JANE_DOE);
+  const currentCard = VALID_VISA_CREDIT;
   const ticketQuantity = 3;
   const eventsPage = new EventsPage(page);
   const doorList = new DoorListPage(page);
