+++ client/e2e/tests/mainTest.spec.ts
@@ -1,15 +1,18 @@
-import {test, expect} from '@playwright/test';
-import {MainPage} from './pages/mainPage';
-import {EventsPage} from './pages/EventsPage';
-import {ContactPage} from './pages/contactPage';
-import {DoorListPage} from './pages/doorListPage';
-import {EventInfoTemplate5, EventInfoTemplate6, ShowingInfo2, ShowingInfo5} from './testData/ConstsPackage';
-import {EventInfo, ValidVisaCredit} from './testData/ConstsPackage';
-import {createUniqueCustomer, baseJohnDoe, baseJaneDoe} from './testData/ConstsPackage';
+import { test, expect } from '@playwright/test';
+import { MainPage } from '../pages/mainPage';
+import { EventsPage } from '../pages/EventsPage';
+import { ContactPage } from '../pages/contactPage';
+import { DoorListPage } from '../pages/doorListPage';
+import { JOHN_DOE, JANE_DOE } from '../test-data/dataConstants/CustomerInfoConstants';
+import { EVENT_INFO_5, EVENT_INFO_6 } from '../test-data/dataConstants/EventInfoConstants';
+import { SHOWING_INFO_2, SHOWING_INFO_5 } from '../test-data/dataConstants/ShowingInfoConstants';
+import { createUniqueEvent, createUniqueCustomer } from '../test-data/factoryFunctions';
+import { VALID_VISA_CREDIT } from '../test-data/dataConstants/CreditCardConstants';
 
 // Verify we can get to the main page and the event header is visible
 test('Check Home', async ({page}) => {
   const main = new MainPage(page);
+
   await main.goto();
   await expect(main.headingEvent).toBeVisible();
 });
@@ -17,6 +20,7 @@ test('Check Home', async ({page}) => {
 // Navigate to the first showing on the main page and make sure the event title isn't blank
 test.skip('view show', async ({page}) => {
   const main = new MainPage(page);
+
   await main.goto();
   const showing = await main.goFirstShowing();
   expect(showing, 'Showing is empty').not.toBeNull();
@@ -25,34 +29,43 @@ test.skip('view show', async ({page}) => {
 // Go through order process and check both the success message and cart contents
 test('check cart after ticket add', async ({page}) => {
   test.setTimeout(60000);
+
   const events = new EventsPage(page);
   const main = new MainPage(page);
-  const currentEvent = new EventInfo(EventInfoTemplate5);
-  const currentShowing = ShowingInfo2;
+
+  const currentEvent = createUniqueEvent(EVENT_INFO_5);
+  const currentShowing = SHOWING_INFO_2;
+
   await events.goto();
   await events.addnewevent(currentEvent);
   await events.activateEvent();
   await events.addNewShowing(currentShowing);
+  
   try {
     await main.goto();
     const showing = await main.goSelectShowing(currentEvent);
     // Rebuild randoms to use a fixed selection using the EventInfo and ShowingInfo
     expect(showing).toEqual(currentEvent.eventName);
+
     const date = await main.selectRandomDate();
     expect(date).toEqual(currentShowing.showingWholeDate);
+
     const time = await main.selectRandomTime();
     const ticketType = await main.selectRandomTicketType();
     const quantity = await main.selectRandomQuantity();
     const dateParts = date.split(' ');
+
     let confirmMessage: string;
     confirmMessage = 'You added ' + quantity + ' ticket';
     if (parseInt(quantity) > 1) {
       confirmMessage += 's';
     }
     confirmMessage += ' to ' + showing + ' on ' + dateParts[1] + ' ' + dateParts[2] + ' to the cart.';
+
     await main.clickGetTickets();
     await main.checkAddTicketSuccess(confirmMessage);
     await main.clickTakeMeThere();
+
     const cartInfo = ticketType + ' - ' + dateParts[1] + ' - ' + time;
     await main.checkCart(currentEvent, cartInfo, quantity);
   } finally {
@@ -66,16 +79,21 @@ test('check cart after ticket add', async ({page}) => {
 test('check stripe purchase', async ({page}, testInfo) => {
   const timeoutAdd = testInfo.retry * 5000;
   test.setTimeout(80000 + (timeoutAdd * 2)); // There are two places with extended timeouts
-  const currentPatron = createUniqueCustomer(baseJohnDoe);
-  const currentCard = ValidVisaCredit;
-  const currentEvent = new EventInfo(EventInfoTemplate5);
-  const currentShowing = ShowingInfo2;
+
   const events = new EventsPage(page);
   const main = new MainPage(page);
+
+  const currentEvent = createUniqueEvent(EVENT_INFO_5);
+  const currentPatron = createUniqueCustomer(JOHN_DOE);
+
+  const currentCard = VALID_VISA_CREDIT;
+  const currentShowing = SHOWING_INFO_2;
+
   await events.goto();
   await events.addnewevent(currentEvent);
   await events.activateEvent();
   await events.addNewShowing(currentShowing);
+
   try {
     await main.goto();
     await main.purchaseTicket(currentPatron, currentCard, currentEvent, {timeoutAdd: timeoutAdd});
@@ -92,14 +110,17 @@ test('check stripe purchase', async ({page}, testInfo) => {
 test('check contact is added after order', async ({page}, testInfo) => {
   const timeoutAdd = testInfo.retry * 5000;
   test.setTimeout(80000);
-  const currentPatron = createUniqueCustomer(baseJohnDoe);
 
-  const currentCard = ValidVisaCredit;
-  const currentEvent = new EventInfo(EventInfoTemplate5);
-  const currentShowing = ShowingInfo2;
   const events = new EventsPage(page);
   const main = new MainPage(page);
   const contacts = new ContactPage(page);
+
+  const currentEvent = createUniqueEvent(EVENT_INFO_5);
+  const currentShowing = SHOWING_INFO_2;
+
+  const currentPatron = createUniqueCustomer(JOHN_DOE);
+  const currentCard = VALID_VISA_CREDIT;
+
   await events.goto();
   await events.addnewevent(currentEvent);
   await events.activateEvent();
@@ -121,14 +142,17 @@ test('check contact is added after order', async ({page}, testInfo) => {
 test('check order accommodations', async ({page}, testInfo) => {
   const timeoutAdd = testInfo.retry * 5000;
   test.setTimeout(80000 + timeoutAdd);
-  const currentPatron = createUniqueCustomer(baseJaneDoe);
 
-  const currentCard = ValidVisaCredit;
-  const currentEvent = new EventInfo(EventInfoTemplate5);
-  const currentShowing = ShowingInfo2;
   const events = new EventsPage(page);
   const main = new MainPage(page);
   const contacts = new ContactPage(page);
+
+  const currentEvent = createUniqueEvent(EVENT_INFO_5);
+  const currentShowing = SHOWING_INFO_2;
+
+  const currentPatron = createUniqueCustomer(JANE_DOE);
+  const currentCard = VALID_VISA_CREDIT;
+
   await events.goto();
   await events.addnewevent(currentEvent);
   await events.activateEvent();
@@ -149,11 +173,15 @@ test('check order accommodations', async ({page}, testInfo) => {
 // Place a ticket in the cart and ensure both the increment and decrement buttons work
 test('check ticket inc/dec in cart', async ({page}) => {
   test.setTimeout(60000);
+  
   const events = new EventsPage(page);
   const main = new MainPage(page);
+
+  const currentEvent = createUniqueEvent(EVENT_INFO_5);
+  const currentShowing = SHOWING_INFO_2;
+
   const quantity = 2;
-  const currentEvent = new EventInfo(EventInfoTemplate5);
-  const currentShowing = ShowingInfo2;
+
   await events.goto();
   await events.addnewevent(currentEvent);
   await events.activateEvent();
@@ -183,15 +211,18 @@ test('check ticket inc/dec in cart', async ({page}) => {
 test('check order on door list', async ({page}, testInfo) => {
   const timeoutAdd = testInfo.retry * 5000;
   test.setTimeout(100000 + timeoutAdd);
-  const currentPatron = createUniqueCustomer(baseJohnDoe);
 
-  const currentCard = ValidVisaCredit;
-  const currentEvent = new EventInfo(EventInfoTemplate6);
-  const currentShowing = ShowingInfo5;
-  const quantity = 2;
   const events = new EventsPage(page);
   const main = new MainPage(page);
   const doorList = new DoorListPage(page);
+
+  const currentEvent = createUniqueEvent(EVENT_INFO_6);
+  const currentShowing = SHOWING_INFO_5;
+
+  const currentPatron = createUniqueCustomer(JOHN_DOE);
+  const currentCard = VALID_VISA_CREDIT;
+  const quantity = 2;
+
   await events.goto();
   await events.addnewevent(currentEvent);
   await events.activateEvent();
